불필요한 코드 추가, 제거
